feat: configuration from september 2024
First config I made, im kind of migrating it to a newer one pc, just to have log of it. 
Currently i have a laptop from Toshiba's Satellite brand, its, if im not wrong, a S552 series, the problem is that it often overheats about 70+C grades easily (https://envs.sh/mtV.png) so u would say probably that i have to clean the fan or change the thermal paste for the processor, but over a year i try it and it is just impossible to unscrew some screws, and there is the risk of broke or damage a piece of the motherboard. Anyways, it has already a decade being used, so its time to change it. Also it doesnt have a graphic board, it just the integrate radeon graphics that it uses to do all of the work, it does that the core's temperature increase. To have a memory of it, it has a AMD A8-4555M (its an APU) with Radeon 7600G (discrete) and 4 cores of 1.5GHz each. 8GB RAM DDR3 and 1TB HDD disk. And a screen of 15'' 60hz 1366x768 of resolution.

Laptops are a scam nowadays, i'll get a workstation instead.